+++ examples/trajectory_reconstruction.py
@@ -1,4 +1,6 @@
 r"""
+.. _trajectory_stride:
+
 Stride Level Trajectory reconstruction
 ======================================
 
@@ -63,7 +65,7 @@ trajectory = StrideLevelTrajectory(ori_method, pos_method)
 sampling_frequency_hz = 204.8
 trajectory.estimate(data=imu_data, stride_event_list=stride_list, sampling_rate_hz=sampling_frequency_hz)
 
-# select the postion of the first stride
+# select the position of the first stride
 first_stride_position = trajectory.position_["left_sensor"].loc[0]
 
 first_stride_position.plot()